backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
requests>=2.31.0
pydantic>=2.9.0
httpx>=0.24.0
openai>=1.40.0
tenacity>=8.2.3
tinydb>=4.8.0
loguru>=0.7.0

# v11.0 Neural Link 身份识别系统依赖
sqlalchemy>=2.0.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
captcha>=0.6.0
